updated Lambda examples
