--- javav2/example_code/lambda/src/main/java/com/example/lambda/CreateFunction.java
@@ -1,111 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.example.lambda;
-
-// snippet-start:[lambda.java2.create.main]
-// snippet-start:[lambda.java2.create.import]
-import software.amazon.awssdk.core.SdkBytes;
-import software.amazon.awssdk.core.waiters.WaiterResponse;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.lambda.LambdaClient;
-import software.amazon.awssdk.services.lambda.model.CreateFunctionRequest;
-import software.amazon.awssdk.services.lambda.model.FunctionCode;
-import software.amazon.awssdk.services.lambda.model.CreateFunctionResponse;
-import software.amazon.awssdk.services.lambda.model.GetFunctionRequest;
-import software.amazon.awssdk.services.lambda.model.GetFunctionResponse;
-import software.amazon.awssdk.services.lambda.model.LambdaException;
-import software.amazon.awssdk.services.lambda.model.Runtime;
-import software.amazon.awssdk.services.lambda.waiters.LambdaWaiter;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.InputStream;
-// snippet-end:[lambda.java2.create.import]
-
-/**
- * This code example requires a ZIP or JAR that represents the code of the
- * Lambda function.
- * If you do not have a ZIP or JAR, please refer to the following document:
- *
- * https://github.com/aws-doc-sdk-examples/tree/master/javav2/usecases/creating_workflows_stepfunctions
- *
- * Also, set up your development environment, including your credentials.
- *
- * For information, see this documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- */
-
-public class CreateFunction {
-    public static void main(String[] args) {
-
-        final String usage = """
-
-                Usage:
-                    <functionName> <filePath> <role> <handler>\s
-
-                Where:
-                    functionName - The name of the Lambda function.\s
-                    filePath - The path to the ZIP or JAR where the code is located.\s
-                    role - The role ARN that has Lambda permissions.\s
-                    handler - The fully qualified method name (for example, example.Handler::handleRequest). \s
-                """;
-
-        if (args.length != 4) {
-            System.out.println(usage);
-            System.exit(1);
-        }
-
-        String functionName = args[0];
-        String filePath = args[1];
-        String role = args[2];
-        String handler = args[3];
-        Region region = Region.US_WEST_2;
-        LambdaClient awsLambda = LambdaClient.builder()
-                .region(region)
-                .build();
-
-        createLambdaFunction(awsLambda, functionName, filePath, role, handler);
-        awsLambda.close();
-    }
-
-    public static void createLambdaFunction(LambdaClient awsLambda,
-            String functionName,
-            String filePath,
-            String role,
-            String handler) {
-
-        try {
-            LambdaWaiter waiter = awsLambda.waiter();
-            InputStream is = new FileInputStream(filePath);
-            SdkBytes fileToUpload = SdkBytes.fromInputStream(is);
-
-            FunctionCode code = FunctionCode.builder()
-                    .zipFile(fileToUpload)
-                    .build();
-
-            CreateFunctionRequest functionRequest = CreateFunctionRequest.builder()
-                    .functionName(functionName)
-                    .description("Created by the Lambda Java API")
-                    .code(code)
-                    .handler(handler)
-                    .runtime(Runtime.JAVA8)
-                    .role(role)
-                    .build();
-
-            // Create a Lambda function using a waiter.
-            CreateFunctionResponse functionResponse = awsLambda.createFunction(functionRequest);
-            GetFunctionRequest getFunctionRequest = GetFunctionRequest.builder()
-                    .functionName(functionName)
-                    .build();
-            WaiterResponse<GetFunctionResponse> waiterResponse = waiter.waitUntilFunctionExists(getFunctionRequest);
-            waiterResponse.matched().response().ifPresent(System.out::println);
-            System.out.println("The function ARN is " + functionResponse.functionArn());
-
-        } catch (LambdaException | FileNotFoundException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-    }
-}
-// snippet-end:[lambda.java2.create.main]
