--- javav2/example_code/lambda/src/main/java/com/example/lambda/LambdaScenario.java
@@ -1,298 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.example.lambda;
-
-// snippet-start:[lambda.javav2.scenario.import]
-import org.json.JSONObject;
-import software.amazon.awssdk.core.SdkBytes;
-import software.amazon.awssdk.core.waiters.WaiterResponse;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.lambda.LambdaClient;
-import software.amazon.awssdk.services.lambda.model.FunctionCode;
-import software.amazon.awssdk.services.lambda.model.CreateFunctionResponse;
-import software.amazon.awssdk.services.lambda.model.GetFunctionRequest;
-import software.amazon.awssdk.services.lambda.model.GetFunctionResponse;
-import software.amazon.awssdk.services.lambda.model.LambdaException;
-import software.amazon.awssdk.services.lambda.model.ListFunctionsResponse;
-import software.amazon.awssdk.services.lambda.model.FunctionConfiguration;
-import software.amazon.awssdk.services.lambda.model.InvokeResponse;
-import software.amazon.awssdk.services.lambda.model.InvokeRequest;
-import software.amazon.awssdk.services.lambda.model.CreateFunctionRequest;
-import software.amazon.awssdk.services.lambda.model.UpdateFunctionCodeRequest;
-import software.amazon.awssdk.services.lambda.model.UpdateFunctionCodeResponse;
-import software.amazon.awssdk.services.lambda.model.GetFunctionConfigurationRequest;
-import software.amazon.awssdk.services.lambda.model.GetFunctionConfigurationResponse;
-import software.amazon.awssdk.services.lambda.model.DeleteFunctionRequest;
-import software.amazon.awssdk.services.lambda.model.UpdateFunctionConfigurationRequest;
-import software.amazon.awssdk.services.lambda.model.Runtime;
-import software.amazon.awssdk.services.lambda.waiters.LambdaWaiter;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.InputStream;
-import java.util.List;
-// snippet-end:[lambda.javav2.scenario.import]
-
-// snippet-start:[lambda.javav2.scenario.main]
-/*
- *  Lambda function names appear as:
- *
- *  arn:aws:lambda:us-west-2:335556666777:function:HelloFunction
- *
- *  To find this value, look at the function in the AWS Management Console.
- *
- *  Before running this Java code example, set up your development environment, including your credentials.
- *
- *  For more information, see this documentation topic:
- *
- *  https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- *
- *  This example performs the following tasks:
- *
- * 1. Creates an AWS Lambda function.
- * 2. Gets a specific AWS Lambda function.
- * 3. Lists all Lambda functions.
- * 4. Invokes a Lambda function.
- * 5. Updates the Lambda function code and invokes it again.
- * 6. Updates a Lambda function's configuration value.
- * 7. Deletes a Lambda function.
- */
-
-public class LambdaScenario {
-    public static final String DASHES = new String(new char[80]).replace("\0", "-");
-
-    public static void main(String[] args) throws InterruptedException {
-        final String usage = """
-
-                Usage:
-                    <functionName> <filePath> <role> <handler> <bucketName> <key>\s
-
-                Where:
-                    functionName - The name of the Lambda function.\s
-                    filePath - The path to the .zip or .jar where the code is located.\s
-                    role - The AWS Identity and Access Management (IAM) service role that has Lambda permissions.\s
-                    handler - The fully qualified method name (for example, example.Handler::handleRequest).\s
-                    bucketName - The Amazon Simple Storage Service (Amazon S3) bucket name that contains the .zip or .jar used to update the Lambda function's code.\s
-                    key - The Amazon S3 key name that represents the .zip or .jar (for example, LambdaHello-1.0-SNAPSHOT.jar).
-                    """;
-
-        if (args.length != 6) {
-            System.out.println(usage);
-            System.exit(1);
-        }
-
-        String functionName = args[0];
-        String filePath = args[1];
-        String role = args[2];
-        String handler = args[3];
-        String bucketName = args[4];
-        String key = args[5];
-
-        Region region = Region.US_WEST_2;
-        LambdaClient awsLambda = LambdaClient.builder()
-                .region(region)
-                .build();
-
-        System.out.println(DASHES);
-        System.out.println("Welcome to the AWS Lambda example scenario.");
-        System.out.println(DASHES);
-
-        System.out.println(DASHES);
-        System.out.println("1. Create an AWS Lambda function.");
-        String funArn = createLambdaFunction(awsLambda, functionName, filePath, role, handler);
-        System.out.println("The AWS Lambda ARN is " + funArn);
-        System.out.println(DASHES);
-
-        System.out.println(DASHES);
-        System.out.println("2. Get the " + functionName + " AWS Lambda function.");
-        getFunction(awsLambda, functionName);
-        System.out.println(DASHES);
-
-        System.out.println(DASHES);
-        System.out.println("3. List all AWS Lambda functions.");
-        listFunctions(awsLambda);
-        System.out.println(DASHES);
-
-        System.out.println(DASHES);
-        System.out.println("4. Invoke the Lambda function.");
-        System.out.println("*** Sleep for 1 min to get Lambda function ready.");
-        Thread.sleep(60000);
-        invokeFunction(awsLambda, functionName);
-        System.out.println(DASHES);
-
-        System.out.println(DASHES);
-        System.out.println("5. Update the Lambda function code and invoke it again.");
-        updateFunctionCode(awsLambda, functionName, bucketName, key);
-        System.out.println("*** Sleep for 1 min to get Lambda function ready.");
-        Thread.sleep(60000);
-        invokeFunction(awsLambda, functionName);
-        System.out.println(DASHES);
-
-        System.out.println(DASHES);
-        System.out.println("6. Update a Lambda function's configuration value.");
-        updateFunctionConfiguration(awsLambda, functionName, handler);
-        System.out.println(DASHES);
-
-        System.out.println(DASHES);
-        System.out.println("7. Delete the AWS Lambda function.");
-        LambdaScenario.deleteLambdaFunction(awsLambda, functionName);
-        System.out.println(DASHES);
-
-        System.out.println(DASHES);
-        System.out.println("The AWS Lambda scenario completed successfully");
-        System.out.println(DASHES);
-        awsLambda.close();
-    }
-
-    public static String createLambdaFunction(LambdaClient awsLambda,
-            String functionName,
-            String filePath,
-            String role,
-            String handler) {
-
-        try {
-            LambdaWaiter waiter = awsLambda.waiter();
-            InputStream is = new FileInputStream(filePath);
-            SdkBytes fileToUpload = SdkBytes.fromInputStream(is);
-
-            FunctionCode code = FunctionCode.builder()
-                    .zipFile(fileToUpload)
-                    .build();
-
-            CreateFunctionRequest functionRequest = CreateFunctionRequest.builder()
-                    .functionName(functionName)
-                    .description("Created by the Lambda Java API")
-                    .code(code)
-                    .handler(handler)
-                    .runtime(Runtime.JAVA8)
-                    .role(role)
-                    .build();
-
-            // Create a Lambda function using a waiter
-            CreateFunctionResponse functionResponse = awsLambda.createFunction(functionRequest);
-            GetFunctionRequest getFunctionRequest = GetFunctionRequest.builder()
-                    .functionName(functionName)
-                    .build();
-            WaiterResponse<GetFunctionResponse> waiterResponse = waiter.waitUntilFunctionExists(getFunctionRequest);
-            waiterResponse.matched().response().ifPresent(System.out::println);
-            return functionResponse.functionArn();
-
-        } catch (LambdaException | FileNotFoundException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-        return "";
-    }
-
-    public static void getFunction(LambdaClient awsLambda, String functionName) {
-        try {
-            GetFunctionRequest functionRequest = GetFunctionRequest.builder()
-                    .functionName(functionName)
-                    .build();
-
-            GetFunctionResponse response = awsLambda.getFunction(functionRequest);
-            System.out.println("The runtime of this Lambda function is " + response.configuration().runtime());
-
-        } catch (LambdaException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-    }
-
-    public static void listFunctions(LambdaClient awsLambda) {
-        try {
-            ListFunctionsResponse functionResult = awsLambda.listFunctions();
-            List<FunctionConfiguration> list = functionResult.functions();
-            for (FunctionConfiguration config : list) {
-                System.out.println("The function name is " + config.functionName());
-            }
-
-        } catch (LambdaException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-    }
-
-    public static void invokeFunction(LambdaClient awsLambda, String functionName) {
-
-        InvokeResponse res;
-        try {
-            // Need a SdkBytes instance for the payload.
-            JSONObject jsonObj = new JSONObject();
-            jsonObj.put("inputValue", "2000");
-            String json = jsonObj.toString();
-            SdkBytes payload = SdkBytes.fromUtf8String(json);
-
-            InvokeRequest request = InvokeRequest.builder()
-                    .functionName(functionName)
-                    .payload(payload)
-                    .build();
-
-            res = awsLambda.invoke(request);
-            String value = res.payload().asUtf8String();
-            System.out.println(value);
-
-        } catch (LambdaException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-    }
-
-    public static void updateFunctionCode(LambdaClient awsLambda, String functionName, String bucketName, String key) {
-        try {
-            LambdaWaiter waiter = awsLambda.waiter();
-            UpdateFunctionCodeRequest functionCodeRequest = UpdateFunctionCodeRequest.builder()
-                    .functionName(functionName)
-                    .publish(true)
-                    .s3Bucket(bucketName)
-                    .s3Key(key)
-                    .build();
-
-            UpdateFunctionCodeResponse response = awsLambda.updateFunctionCode(functionCodeRequest);
-            GetFunctionConfigurationRequest getFunctionConfigRequest = GetFunctionConfigurationRequest.builder()
-                    .functionName(functionName)
-                    .build();
-
-            WaiterResponse<GetFunctionConfigurationResponse> waiterResponse = waiter
-                    .waitUntilFunctionUpdated(getFunctionConfigRequest);
-            waiterResponse.matched().response().ifPresent(System.out::println);
-            System.out.println("The last modified value is " + response.lastModified());
-
-        } catch (LambdaException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-    }
-
-    public static void updateFunctionConfiguration(LambdaClient awsLambda, String functionName, String handler) {
-        try {
-            UpdateFunctionConfigurationRequest configurationRequest = UpdateFunctionConfigurationRequest.builder()
-                    .functionName(functionName)
-                    .handler(handler)
-                    .runtime(Runtime.JAVA11)
-                    .build();
-
-            awsLambda.updateFunctionConfiguration(configurationRequest);
-
-        } catch (LambdaException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-    }
-
-    public static void deleteLambdaFunction(LambdaClient awsLambda, String functionName) {
-        try {
-            DeleteFunctionRequest request = DeleteFunctionRequest.builder()
-                    .functionName(functionName)
-                    .build();
-
-            awsLambda.deleteFunction(request);
-            System.out.println("The " + functionName + " function was deleted");
-
-        } catch (LambdaException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-    }
-}
-// snippet-end:[lambda.javav2.scenario.main]
