--- javav2/example_code/lambda/src/main/java/com/example/lambda/ListLambdaFunctions.java
@@ -1,51 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-// snippet-start:[lambda.java2.ListLambdaFunctions.complete]
-package com.example.lambda;
-
-// snippet-start:[lambda.java2.list.main]
-// snippet-start:[lambda.java2.list.import]
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.lambda.LambdaClient;
-import software.amazon.awssdk.services.lambda.model.LambdaException;
-import software.amazon.awssdk.services.lambda.model.ListFunctionsResponse;
-import software.amazon.awssdk.services.lambda.model.FunctionConfiguration;
-import java.util.List;
-// snippet-end:[lambda.java2.list.import]
-
-/**
- * Before running this Java V2 code example, set up your development
- * environment, including your credentials.
- *
- * For more information, see the following documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- */
-public class ListLambdaFunctions {
-    public static void main(String[] args) {
-        Region region = Region.US_WEST_2;
-        LambdaClient awsLambda = LambdaClient.builder()
-                .region(region)
-                .build();
-
-        listFunctions(awsLambda);
-        awsLambda.close();
-    }
-
-    public static void listFunctions(LambdaClient awsLambda) {
-        try {
-            ListFunctionsResponse functionResult = awsLambda.listFunctions();
-            List<FunctionConfiguration> list = functionResult.functions();
-            for (FunctionConfiguration config : list) {
-                System.out.println("The function name is " + config.functionName());
-            }
-
-        } catch (LambdaException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-    }
-}
-// snippet-end:[lambda.java2.list.main]
-// snippet-end:[lambda.java2.ListLambdaFunctions.complete]
